Updated implementation using nested dicts.
The previous implementation relied on the (later debunked) assumption
that calling rcl_parse_yaml_file repeatedly with the same rcl_params_t
would yield a struct that contains the cumulative parameters from those
files. That isn't actually the case so we need to create an intermediate
storage which can be used instead. In rclcpp a std::map is used but we
have no such datatype here. Instead we'll use a PyDict which maps node
names to a nested dict of rclpy.parameter.Parameters by parameter name
and then query that struct for parameters to return to the node.