Remove __eq__ and __hash__ from Subscription
Implementations were using the address held by the pycapsule as the
hash.
When the subscription is destroyed that memory address is freed.
This makes it possible for a new subscription to be created using the
same address, and therefor having the same hash.

This is a problem in the unit tests that reuse the same node for
multiple tests.
After the subscription is destroyed in one test, a new test can create a
subscription that has the same memory address.
If the garbage collector has not cleaned up the old subscription object
then when the new subscription is added to the node's callback group it
will have an identical hash to the destroyed subscription.
Then after the garbage collector runs the callback group's weakref to
the old subscription is deleted.
This leads to an assertion error in callback_group.can_execute because
there is no weakref to the newly created subscription.

Signed-off-by: Shane Loretz <sloretz@osrfoundation.org>