Prevent rcutils_log from accessing invalid memory
Before this fix, rclpy_logging_rcutils_log was using
rcutils_log without a format string (log message was
passed directly as the foramt).

This is unsafe because if the log message contains
a sequence of character which can be interpreted by
snprintf ("%d" for instance), snprintf will look for
this argument value which was never passed.
This leads to reading invalid memory and may cause a crash
or corrupted log messages.

The simple fix is to always pass "%s" when a static log
string needs to be logged.

Signed-off-by: Thomas Moulard <tmoulard@amazon.com>