Remove most of the timing checks in test_executor
The threaded sleep waits *at least* 0.1 seconds from when it
is started, but there are no guarantees on when the other
thread gets scheduled in to finish the testing.  Thus, just
remove the various timing checks and just make sure that the
future completes; if it doesn't, it will get picked up by
the overall test timeout.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>