Fix Python 3.5 compatibility by removing unsupported type hints.
Some of the type hints introduced in Dashing are not compatible with
Python 3.5. Where possible I tried to convert the hints to type comments
but for the type hints within function definitions that did not appear
to be possible so those hints were simply removed.

This is going to make backporting changes more challenging since we'll
have to:
1. Review backported changes for incompatible type hints.
2. Reconcile merge conflicts caused by the removal of type hints on the
dashing development branches.

Signed-off-by: Steven! Ragnarök <steven@nuclearsandwich.com>