[Dashing] Fix Python 3.5 compatibility by removing unsupported type hints. (#414)
* Fix Python 3.5 compatibility by removing unsupported type hints.

Some of the type hints introduced in Dashing are not compatible with
Python 3.5. Where possible I tried to convert the hints to type comments
but for the type hints within function definitions that did not appear
to be possible so those hints were simply removed.

This is going to make backporting changes more challenging since we'll
have to:
1. Review backported changes for incompatible type hints.
2. Reconcile merge conflicts caused by the removal of type hints on the
dashing development branches.

* Restore type hints in parameter lists.

Also re-adds imports that are used in hints.
The trailing commas on some parameter lists came up as syntax errors.

* Don't warn on imports for type comments.

According to [PEP-0484] types used in type comments must be imported or
defined in the current module. Since flake8 is not type-comment aware
(at least on Dashing PR jobs) I've noqa'd the imports.

[PEP-0484]: http://www.python.org/dev/peps/pep-0484/

Signed-off-by: Steven! Ragnarök <steven@nuclearsandwich.com>