Fix two clang static analysis warnings.
The first one in pycapsule is truly a bug; if destructor
was ever NULL, we would have reported an error and then
immediately dereferenced the NULL pointer, leading to a
crash.

The second one is a false positive.  The documentation for
PyErr_NoMemory() says that it always returns NULL, but
clang static analysis can't tell that just from the function
declaration.  Thus, call PyErr_NoMemory() to set the error,
and return NULL ourselves.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>