Properly implement action server/client handle cleanup.
In particular, it should never really be the case that the
underlying handle that we are using is None.  That only seemed
to be happening because we were double destroying; once in
the explicit call to destroy(), and once in __del__.  But
we don't actually need __del__; during garbage collection, we'll
drop the reference to the handle and then the underlying object
will get freed anyway.  Just remove all of that extraneous
infrastructure here.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>