Add in a warning for a depth of 0 with KEEP_LAST.
It really doesn't make much sense to have a KeepLast depth of 0;
no data could possibly be stored. Indeed, the underlying DDS
implementations don't actually support this. It currently "works"
because a KeepLast depth of 0 is assumed to be system default, so
the RMW layer typically chooses "1" instead. But this isn't
something we should be encouraging users to do, so add a warning.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>