Don't crash the action server if the client goes away. (#1114)
While doing some testing locally, I had a broken action
client that would send a goal, and then crash.  That
meant it wasn't available to receive the goal response,
nor the result response.  In this situation, my action
server was throwing an exception and crashing.

But that doesn't seem right.  The action server should continue
to run so other action clients can connect.  There are at
least 2 different ways we could accomplish this:

1. Document that the user needs to catch these exceptions
and deal with it themselves.
2. Handle these exceptions at the ActionServer layer.

There is some benefit to 1, in that the user has full control.
However, the problem is that the exceptions we throw in rclpy
aren't detailed enough for the user to disambiguate the errors.
So instead, we go for option 2 here.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>