Add in the ability to start timers paused.
This was possible in the past by creating a timer
and then immediately canceling it.  But that could
lead to a potential race where the timer fired once
before it was canceled.  By allowing for the user
to create it paused, we avoid that race.

This is the rclpy version of https://github.com/ros2/rclcpp/pull/2005

Signed-off-by: Chris Lalancette <clalancette@gmail.com>