Make rclpy initialization context-manager aware.
This PR does two somewhat controversial things:

1.  It adds in a new "rclpy.managed_init" method to rclpy,
which would be used with a context manager instead of
rclpy.init.  I had to do things this way because there
is no way (as far as I know) for the callee to know whether
it was called within a context manager or not, and I
don't think we can reasonably change the semantics of
rclpy.init() at this point.

2.  It changes the context manager implementation of
Context so that it does *not* call init() within itself.
This definitely changes the semantics, but as it stands
that initialization doesn't make sense because it can't
take arguments.  I think this change is warranted,
though we may have to search through documentation and
examples to make sure this doesn't break anything.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>