Add top-level try_shutdown method.
I noticed that this was needed when trying to convert some
of the examples over to the context manager.  They currently
throw an exception because the signal handler does the
shutdown, and then leaving the context tries to do it
again.

This is also more symmetric with what shutdown() does,
in that it also uninstalls signal handlers.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>