Fix a bug when using multiple rclpy.init context managers. (#1314)
That is, if you call rclpy.init as a context manager within
the same process, and using the default context, it currently
fails.  That is because you cannot reinitialize (call init) on
a Context object once it has been shutdown.  Instead, force
try_shutdown to replace the global context object with a new
one on rclpy.init shutdown by passing a None.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>